src/main/java/com/example/demo/customer/services/SearchCustomersByNameService.java
package com.example.demo.customer.services;

import com.example.demo.customer.dtos.CustomerDto;
import com.example.demo.customer.repositories.CustomerRepository;
import java.util.List;
import java.util.stream.Collectors;
import lombok.NonNull;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

@Service
@RequiredArgsConstructor
public class SearchCustomersByNameService {

    private final CustomerRepository customerRepository;
    private final CustomerMapper customerMapper;

    @Transactional(readOnly = true)
    public List<CustomerDto> searchCustomersByName(@NonNull String name) {
        return customerRepository.findByName(name).stream()
                .map(customerMapper::map)
                .collect(Collectors.toList());
    }
}
